# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Copy Trading Admin Panel API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /trade/execute:
    post:
      operationId: executeTrade
      tags: [trade]
      summary: Execute trade across accounts
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TradeExecution"
      responses:
        "200":
          description: Per-account results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AccountResult"
  /trade/order:
    delete:
      operationId: cancelOrder
      tags: [trade]
      summary: Cancel a specific order across accounts
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CancelOrderInput"
      responses:
        "200":
          description: Cancel results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AccountResult"
  /trade/cancel-all:
    post:
      operationId: cancelAllOrders
      tags: [trade]
      summary: Cancel all open orders
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CancelAllInput"
      responses:
        "200":
          description: Cancel all results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AccountResult"
...
  /orders/open:
    post:
      operationId: getOpenOrders
      tags: [orders]
      summary: Get open orders across accounts
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpenOrdersQuery"
      responses:
        "200":
          description: Open orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrderWithAccount"
  /orders/closed:
    post:
      operationId: getClosedOrders
      tags: [orders]
      summary: Get closed orders
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClosedOrdersQuery"
      responses:
        "200":
          description: Closed orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrderWithAccount"
...
  /tpsl/set:
    post:
      operationId: setTpsl
      tags: [tpsl]
      summary: Set TP/SL for accounts
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TpslInput"
      responses:
        "200":
          description: Results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AccountResult"
  /tpsl/{accountId}/{orderId}:
    delete:
      operationId: cancelTpsl
      tags: [tpsl]
      summary: Cancel a TP/SL order
      security:
        - bearerAuth: []
      parameters:
        - name: accountId
          in: path
          required: true
          schema:
            type: integer
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Cancelled
...
  /market/ticker:
    get:
      operationId: getTicker
      tags: [market]
      summary: Get ticker for a symbol
      security:
        - bearerAuth: []
      parameters:
        - name: symbol
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Ticker data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TickerData"
  /market/orderbook:
    get:
      operationId: getOrderbook
      tags: [market]
      summary: Get order book for a symbol
      security:
        - bearerAuth: []
      parameters:
        - name: symbol
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Orderbook
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderbookData"
  /market/instruments:
    get:
      operationId: getInstruments
      tags: [market]
      summary: Get all tradeable instruments
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Instruments map
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InstrumentsData"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    OrderPayload:
      type: object
      required: [symbol, side, orderType, quantity]
      properties:
        symbol:
          type: string
        side:
          type: string
          enum: [BUY, SELL]
        orderType:
          type: string
          enum: [MARKET, LIMIT, STOP_MARKET, TAKE_PROFIT_MARKET]
        quantity:
          type: number
        price:
          type: ["number", "null"]
        triggerPrice:
          type: ["number", "null"]
        reduceOnly:
          type: ["boolean", "null"]
        timeInForce:
          type: ["string", "null"]
    TradeExecution:
      type: object
      required: [accountIds, order]
      properties:
        accountIds:
          type: array
          items:
            type: integer
        order:
          $ref: "#/components/schemas/OrderPayload"
    AccountResult:
      type: object
      required: [accountId, accountName, success]
      properties:
        accountId:
          type: integer
        accountName:
          type: string
        success:
          type: boolean
        orderId:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
        error:
          type: ["string", "null"]
    CancelOrderInput:
      type: object
      required: [accountIds, orderId]
      properties:
        accountIds:
          type: array
          items:
            type: integer
        orderId:
          type: string
    CancelAllInput:
      type: object
      required: [accountIds]
      properties:
        accountIds:
          type: array
          items:
            type: integer
        symbol:
          type: ["string", "null"]
...
    OpenOrdersQuery:
      type: object
      properties:
        symbol:
          type: ["string", "null"]
        accountIds:
          type: array
          items:
            type: integer
    ClosedOrdersQuery:
      type: object
      properties:
        symbol:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
        accountIds:
          type: array
          items:
            type: integer
        fromTime:
          type: ["integer", "null"]
        toTime:
          type: ["integer", "null"]
    OrderWithAccount:
      type: object
      required: [accountId, accountName, orderId, symbol, side, orderType, quantity, status]
      properties:
        accountId:
          type: integer
        accountName:
          type: string
        orderId:
          type: string
        symbol:
          type: string
        side:
          type: string
        orderType:
          type: string
        quantity:
          type: string
        execQuantity:
          type: string
        price:
          type: string
        avgExecutionPrice:
          type: ["string", "null"]
        executionFee:
          type: ["string", "null"]
        realisedPnl:
          type: ["string", "null"]
        reduceOnly:
          type: boolean
        status:
          type: string
        createdAt:
          type: ["integer", "null"]
...
    TpslInput:
      type: object
      required: [accountIds, symbol]
      properties:
        accountIds:
          type: array
          items:
            type: integer
        symbol:
          type: string
        tpPrice:
          type: ["number", "null"]
        slPrice:
          type: ["number", "null"]
...
    TickerData:
      type: object
      required: [symbol]
      properties:
        symbol:
          type: string
        lastPrice:
          type: string
        markPrice:
          type: ["number", "null"]
        indexPrice:
          type: ["number", "null"]
        fundingRate:
          type: ["number", "null"]
        bestBidPrice:
          type: string
        bestAskPrice:
          type: string
        high24h:
          type: string
        low24h:
          type: string
        priceChangePct24h:
          type: string
    OrderbookData:
      type: object
      required: [symbol, bids, asks]
      properties:
        symbol:
          type: string
        bids:
          type: array
          items:
            type: array
            items:
              type: string
        asks:
          type: array
          items:
            type: array
            items:
              type: string
        timestamp:
          type: ["integer", "null"]
    InstrumentsData:
      type: object
      properties:
        instruments:
          type: object
...