# lib/api-zod/src/generated/types/orderWithAccount.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Copy Trading Admin Panel API
 * OpenAPI spec version: 0.1.0
 */

export interface OrderWithAccount {
  accountId: number;
  accountName: string;
  orderId: string;
  symbol: string;
  side: string;
  orderType: string;
  quantity: string;
  execQuantity?: string;
  price?: string;
  /** @nullable */
  avgExecutionPrice?: string | null;
  /** @nullable */
  executionFee?: string | null;
  /** @nullable */
  realisedPnl?: string | null;
  reduceOnly?: boolean;
  status: string;
  /** @nullable */
  createdAt?: number | null;
}